feat(#564): add global keyboard shortcut K to open mood log modal
- Extracted LogEntryForm logic from LogFormPage
- Created useGlobalShortcut hook that prevents triggering while user is typing in inputs
- Built lightweight MoodLogModal wrapper around the form
- Wired modal and shortcut into AppShell and added the requested K navigation hint
- Fixed duplicate ToastProvider import in providers.tsx
