+++ frontend/src/app/providers.tsx
@@ -1,7 +1,6 @@
 import { type ReactNode } from 'react'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { BrowserRouter } from 'react-router-dom'
-import { ToastProvider } from '../components/toast-provider'
 import { AuthProvider } from '../lib/auth-context'
 import { ToastProvider } from '../lib/use-toast'
 import { queryClient } from '../lib/query-client'
